Create SampleOptim
The SampleOptim R-toolbox aims to provide quality indicators estimations to support user’s decision to determine the optimal number sample size for age-length keys (ALKs) and maturity ogives (MO). The simulation process works at sample level. The dataset used should represents the “whole” available data, and the simulations are based on subsampling without replacement by default.
The subsamples obtained from the simulations intend to allow the comparison of the ALKs and of the maturity ogive parameters based on a reduction on the number of individuals sampled by length class.
The setup for simulations allows the user to perform different types of scenarios taking into account a length stratification together with other possible stratifications:
-	temporal stratification (annual, semester and quarter);
-	sex stratification, by defining the sexratio on the subsamples (i.e. proportion of females and males by length class);
-	port stratification (harbour of the samples provenience), options are using a randomly sampling by port, or define a uniform sample distribution by port.
The main aim of SampleOptim is to evaluate a reduction in the number of samples collected to estimate the biological parameters used on the assessment. Although, in cases were the actual sample size is not enough, the functions in the algorithms will not converge which could indicate some bias on the original data sampling.